feature(reports): support process reports grouped by none distributed by process
relates to OPT-5518

+++ backend/src/it/java/org/camunda/optimize/service/es/report/process/single/processinstance/duration/groupby/none/ProcessInstanceDurationByNoneReportEvaluationIT.java
@@ -239,7 +239,7 @@ public class ProcessInstanceDurationByNoneReportEvaluationIT extends AbstractPro
     ReportResultResponseDto<Double> result = reportClient.evaluateNumberReport(reportData).getResult();
 
     // then
-    assertThat(result.getInstanceCount()).isEqualTo((long) selectedTenants.size());
+    assertThat(result.getInstanceCount()).isEqualTo(selectedTenants.size());
   }
 
   @Test
